Move `RecordChunk` under `pieces` module and minor cleanup
