+++ subspace/crates/subspace-farmer-components/src/plotting.rs
@@ -26,11 +26,11 @@ use std::sync::atomic::{AtomicBool, Ordering};
 use std::sync::Arc;
 use std::time::Duration;
 use subspace_core_primitives::hashes::{blake3_hash, blake3_hash_parallel, Blake3Hash};
-use subspace_core_primitives::pieces::{Piece, PieceIndex, PieceOffset, Record};
+use subspace_core_primitives::pieces::{Piece, PieceIndex, PieceOffset, Record, RecordChunk};
 use subspace_core_primitives::pos::PosSeed;
 use subspace_core_primitives::sectors::{SBucket, SectorId, SectorIndex};
 use subspace_core_primitives::segments::HistorySize;
-use subspace_core_primitives::{PublicKey, RecordChunk};
+use subspace_core_primitives::PublicKey;
 use subspace_data_retrieval::piece_getter::PieceGetter;
 use subspace_erasure_coding::ErasureCoding;
 use subspace_proof_of_space::{Table, TableGenerator};
