+++ subspace/crates/subspace-farmer-components/src/proving.rs
@@ -16,11 +16,11 @@ use ab_merkle_tree::balanced_hashed::BalancedHashedMerkleTree;
 use futures::FutureExt;
 use std::collections::VecDeque;
 use std::io;
-use subspace_core_primitives::pieces::{PieceOffset, Record};
+use subspace_core_primitives::pieces::{PieceOffset, Record, RecordChunk};
 use subspace_core_primitives::pos::PosSeed;
 use subspace_core_primitives::sectors::{SBucket, SectorId};
 use subspace_core_primitives::solutions::{ChunkWitness, Solution, SolutionRange};
-use subspace_core_primitives::{PublicKey, RecordChunk};
+use subspace_core_primitives::PublicKey;
 use subspace_erasure_coding::ErasureCoding;
 use subspace_proof_of_space::Table;
 use thiserror::Error;
