+++ subspace/crates/subspace-farmer-components/src/reading.rs
@@ -18,9 +18,8 @@ use std::simd::Simd;
 use std::str::FromStr;
 use std::{fmt, io};
 use subspace_core_primitives::hashes::blake3_hash;
-use subspace_core_primitives::pieces::{Piece, PieceOffset, Record};
+use subspace_core_primitives::pieces::{Piece, PieceOffset, Record, RecordChunk};
 use subspace_core_primitives::sectors::{SBucket, SectorId};
-use subspace_core_primitives::RecordChunk;
 use subspace_erasure_coding::{ErasureCoding, RecoveryShardState};
 use subspace_proof_of_space::{Table, TableGenerator};
 use thiserror::Error;
